Add __init__.py to flag ldm/ as a package
latent-diffusion has a setup.py and uses `find_packages()`, but since
a few of the directories are missing __init__.py nothing is actaully
installed if you use `pip install ...` on it.  This commit adds the
missing __init__.py files, so it can be installed by pip without a
local checkout.